Move some tests to a new test class
